Convert setupDB from a callback function to a promise function
This is done by wrapping the code for setupDB in a promise constructor.
Now, once we have the database ready, either through it already being
initialized, or through the database request's onsuccess handler,
rather than run a callback function, we have the promise resolve. We
can now use setupDB in a promise chain like:

setupDB('').then(/* function to run after we know the DB is open */)