Make addStickyNote promise-based
Following the same pattern as setupDB, addStickyNote's body is wrapped
in the function for the promise constructor; if the transaction
succeeds and its oncomplete event fires, then we resolve. If it fails,
then we reject. This is seen in action in the submitNote function in
page.js; instead of passing getAndDisplayNotes into addStickyNote as
its callback, we pass it into addStickyNote().then so that it runs
after addStickyNote completes.