Convert getNotes to a promise-based function
Once again, we are following the same pattern as before of wrapping our
code in a promise. Unlike in setupDB and addStickyNote, the promise in
getNotes resolves with some data to pass to the promise's then, and
that data is the list of retrieved notes. Therefore, in the
getAndDisplayNotes function, those notes are used in the callback
passed to getNotes().then.